worked on cookies and logout route handler

+++ src/app/api/users/profile/[id]/route.ts
@@ -1,7 +1,6 @@
 import { NextResponse, NextRequest } from 'next/server';
 import prisma from '@/utils/db';
-import jwt from 'jsonwebtoken';
-import { JWTPayload } from '@/utils/types';
+import { verifyToken } from '@/utils/verifyToken';
 
 interface Props {
     params: { id: string };
@@ -13,37 +12,28 @@ interface Props {
  *  @desc    Delete Profile
  *  @access  private (only user himself can delete his account)
  */
-export async function DELETE(request: NextRequest, { params } : Props) {
+export async function DELETE(request: NextRequest, { params }: Props) {
     try {
-        const user = await prisma.user.findUnique({ where: {id: parseInt(params.id)}});
-        if(!user) {
+        const user = await prisma.user.findUnique({ where: { id: parseInt(params.id) } });
+        if (!user) {
             return NextResponse.json(
                 { message: 'user not found' },
                 { status: 404 }
             );
         }
 
-        const authToken = request.headers.get('authToken') as string;
-        if(!authToken) {
-            return NextResponse.json(
-                { message: 'not token provided, access denied' },
-                { status: 401 } // Unauthorized
-            );
-        }
-
-        const userFromToken = jwt.verify(authToken, process.env.JWT_SECRET as string) as JWTPayload;
-        
-        if(userFromToken.id === user.id) {
-            await prisma.user.delete({ where: { id: parseInt(params.id) }});
+        const userFromToken = verifyToken(request);
+        if (userFromToken !== null && userFromToken.id === user.id) {
+            await prisma.user.delete({ where: { id: parseInt(params.id) } });
             return NextResponse.json(
                 { message: 'your profile (account) has been deleted' },
                 { status: 200 }
             );
         }
-        
+
         return NextResponse.json(
             { message: 'only user himself can delete his profile, forbidden' },
-            { status: 403 } // forbidden
+            { status: 403 }
         )
 
     } catch (error) {
@@ -53,3 +43,6 @@ export async function DELETE(request: NextRequest, { params } : Props) {
         )
     }
 }
+
+// GET
+// PUT
